로그인까지 완료

+++ src/components/LogModi.js
@@ -5,14 +5,19 @@ import { useDispatch, useSelector } from 'react-redux';
 import { logout } from '../_reducer/action'
 import axios from 'axios';
 
-function LogModi(props){
+function LogModi(){
     const [id, setId] = useState(useSelector((state) => state.id));
-    const [password, setPassword] = useState(useSelector((state) => state.password));
     const [name, setName] = useState(useSelector((state) => state.username));
+    const [password, setPassword] = useState('');
     const dispatch = useDispatch();
 
-
     const submitbtn = async () => {
+        if (!password) {
+            return alert("비밀번호를 입력하세요.");
+        }
+        else if (!name) {
+            return alert("이름을 입력하세요.");
+        }
         alert('다시 로그인해주세요')
         window.location.href = '/'
         dispatch(logout())
@@ -23,10 +28,14 @@ function LogModi(props){
     }
 
     const deletebtn = async () => {
+        if (!password) {
+            return alert("비밀번호를 입력하세요.");
+        }
+
         alert('회원이 삭제되었습니다')
         window.location.href = '/'
         dispatch(logout())
-        
+
         const result = await axios.delete('/id', {data : {id}})
         console.log(result)   
     }
@@ -39,7 +48,7 @@ function LogModi(props){
                 <span>ID : </span><input type='text' value={id} readOnly />
                 </div>
                 <div className='modi-pass'>
-                <span>password : </span><input type='password' value={password} onChange={(e) => setPassword(e.target.value)} />
+                <span>password : </span><input type='password' value={password} onChange={(e) => setPassword(e.target.value)}  />
                 </div>
                 <div className='modi-name'>
                 <span>NickName : </span><input type='text' value={name} onChange={(e) => setName(e.target.value)} />
@@ -49,7 +58,7 @@ function LogModi(props){
                 <button onClick={deletebtn}>삭제</button>
                 </div>
                 <hr />
-                <div className='modi-txt'>회원정보를 수정해주세요</div>
+                <div className='modi-txt'>삭제시, 비밀번호를 입력해야합니다.</div>
             </div>
         </div>
     )
